feat(sns-cli): Cleanup store canister after upgrading an SNS-controlled canister (#3738)
This PR prepares for adding a new subcommand to sns-cli, so users who
proposed to upgrade an SNS-controlled canister can easily request the
unused cycles to be refunded to the original Cycles Ledger account that
they used for uploading and storing the chunked Wasm in a store
canister.

The corresponding integration test was updated to exercise the happy
scenario.

For now, the subcommand does not refund cycles, but simply deletes the
store canister. Refunding will be implemented in a follow-up PR.