chore(ckbtc): Separate event types for mainnet and testnet logs (#3720)
Update the mainnet event logs used in tests by downloading from the
mainnet minter canister, which is using the new event type that includes
timestamp.

But the testnet canister is still stuck and likely will be re-installed,
the previous downloaded event log for testnet is kept unchanged.

This creates a situation where two event logs are of different formats,
which is addressed by turning `GetEventsFile` into a trait and different
formats then become associated types of the trait.