feat(nervous-system): release runscript automatically creates changelog PRs (#3806)
This is the final step in the release process. Now everything is
automated!

Here is an example of a PR it creates: #3807

To create the PR, we use `gh`, which we assert you have on your computer
before the first step. (This ensures that you don't get to the last step
and then hit an error which would be slightly annoying to recover from.
Instead, you get the error right away, and it should be smooth sailing
after that.)

The intention of this change is that you create the PR to update the
changelog right when you create the proposals.

> Yes, we can optimistically assume that the proposal will pass. If it
doesn't, we can revert the CHANGELOG.md. Branch predict FTW.
- @daniel-wong-dfinity-org 

A consequence of is that the changelog now shows the date of proposal
submission, rather than execution.

[← Previous PR](https://github.com/dfinity/ic/pull/3796)