chore(IDX): don't store bazel-targets (#3851)
This drops the `bazel-targets` which we previously used to keep a list
of built targets. The list was used in `build-determinism` to check
which targets were built by the `bazel-test-all` job, but since both
`build-ic` and `bazel-test-all` use the same diffing logic (`diff.sh`)
it's not necessary to pass the list around.

Also includes a necessary fix that ensure that even if there are no
targets, `build-ic.tar` is a valid tarball.

---------

Co-authored-by: IDX GitHub Automation <idx@dfinity.org>