feat: [MR-353]  Asynchronously mark checkpoint readonly and sync (#3682)
Currently, during checkpointing, we block until we sync the entire
checkpoint and have it moved to the checkpoints folder.
This step can be done asynchronously. 
Moreover, due to the asynchronously writing of pbuf files, this step
should be moved after that and right before removing the marker for
unverified checkpoint.