chore(nns): Change NNS Root code location to one level up (#3873)
The `CHANGELOG.md` and `unreleased_changelog.md` are located in
`rs/nns/handlers/root`, and the `code_location_...` is used to figure
out where the change logs are. Although one obvious solution is to move
the change logs to `//rs/nns/handlers/root/impl/`, there doesn't seem to
be a good reason that changes in `rs/nns/handlers/root/interface` should
NOT be displayed in upgrade proposals.