feat: CON-1422 Implement the VetKd payload builder and validator (#3613)
This PR implements the VetKd payload builder and validator. 

VetKd requests in the state are answered by executing the following
operations:
1. Collect all VetKd shares in the pool for which no agreement has been
created yet
2. Iterate over all VetKd requests without agreement
3. Reject the request if it is expired or requesting an invalid key ID
4. Otherwise, combine VetKd shares created for this request and include
a success agreement.

We will stop adding more agreements to the payload as soon as the next
candidate no longer fits into the payload size limit.

Received block proposals are declared invalid, if they contain a VetKD
agreement that:
1. was already delivered
2. does not correspond to an open VetKD request
3. was rejected for the wrong reason
4. wasn't rejected when it should have been
5. contains a success response that cannot be decoded
6. the combined key is invalid according to crypto.

TODOs to complete in subsequent PRs:
- Implement unit tests
- Add metrics and logging
- Call crypto endpoints to create/validated combined shares