fix: bitcoin adapter: prune blocks if response empty (#3775)
Before, blocks were only pruned if they were: either processed by the
canister, or below (less height) the anchor.

After, in addition to the above, all blocks are also pruned if the
response contained no blocks. The reason why this is safe is because
during the BFS traversal, we try to return at least one cached block
that is reachable from the anchor. If no block was found, this indicates
that no block in the cache can be reached from the anchor, so they can
be safely discarded.

Secondly, this is also needed because the cache can get full of blocks
which are above the height of the anchor, but can't be served currently
(there is for example a block that's missing between them and the
anchor). And if the cache is full of blocks, the adapter won't be able
to request any more new blocks.