fix(IDX): remove CI github email (#3837)
For any auto-generated commits we need to add both a user and an email,
though this user is not actually logged in. This can cause confusion if
the email address is associated with a real user that doesn't have
anything to do with this commit.

The ideal solution would be to set `github-actions` as the commit
author, but I have not figured out a way to do this. For this reason, we
just leave the email blank.

---------

Co-authored-by: GitHub Actions Bot <>