fix: bitcoin adapter: don't retry aggressively  (#3774)
Before, blocks that timed out were retried with higher priority than
trying out new blocks. Even if this meant overloading the cache.

After, new blocks take priority of retrying older blocks, and no new
block requests are sent if the cache is full.

The reason why this was done in the past was because, without it, there
was a chance of:
- requesting a good block (mainchain)
- the peer time outs for some reason
- the cache gets full with bad blocks in the meantime
- the adapter stalls because the cache is full, and can't retry to
download the good block.
The solution was to directly overload the cache if a new block is to be
retried.

This lead to an issue on testnet4 which contains many forks:
- All peers get saturated with forked "get blocks" requests
- They all time out (this is the default behaviour if peers are asked
for blocks they don't store / are not in the mainchain)
- The adapter will prioritise retrying them, never reaching new blocks
to retry, and stalls.

The solution is to check new blocks first. The reason why the previous
issue doesn't happen anymore (cache full + one timed out good block) is
the cache gets periodically pruned if it only contains unreachable
blocks, in https://github.com/dfinity/ic/pull/3775.

Secondly, 
Before, even if a request for a block timed out, it was still stored in
the getdata_request_info, until a new request was tried for the same
block. This could lead to (on testnet4) all peers being saturated with
requests of bad blocks, timing out. However, because
getdata_request_info was not cleaned up, the num_requests_to_be_sent was
0 for all peers, meaning that the adapter thought that there are 8
(maximum) in flight requests for each peer, and therefore did not
attempt to send any more requests, stalling.

After, after a requests times out, it's corresponding
getdata_request_info entry is removed, allowing the canister to try to
request that peer for a block again.