fix(registry): Update the correct node operator ID in node removal (#3924)
### Summary

This update refines the logic within the `do_remove_node_directly`
function to update the correct node operator ID record.

In the past the `caller_id` and the `node_operator_id` for the node were
always the same. However, since #3285 the `caller_id` and the
`node_operator_id` for the removed node may differ, and this introduces
a bug in this edge case.

The wrong node operator record was retrieved (of the caller), but then
written into the correct record id (of the removed node's node
operator), resulting in lost data when using this new capability.

The bug resulted in a node reward discrepancy for a few operator
records, identified in regular administrative checks and described [in
the
forum](https://forum.dfinity.org/t/issue-with-node-provider-rewards/41109/3)
and mitigated with a few NNS proposals mentioned in the forum thread.

### Changelog

- **Functionality**
- Use the correct `node_operator_id` when incrementing node allowance in
`do_remove_node_directly.rs`.

- **Tests** adjustments

---------

Co-authored-by: Max Summe <maximilian.summe@dfinity.org>