fix(registry): disable replacement of nodes that are active in subnets (#3935)
Direct node replacements of nodes that are active in a subnet may result
in unexpected behavior and potential problems in the current Consensus
code.
So to be on the safe side we need to disable the functionality on the
Registry side until the rest of the core protocol can handle it safely.

---------

Co-authored-by: CI Automation <infra+github-automation@dfinity.org>
Co-authored-by: Arshavir Ter-Gabrielyan <arshavir.ter.gabrielyan@dfinity.org>