refactor: Check wasm execution output only in sandbox (#3728)
This PR refactors the way that state modifications flow from the sandbox
where canister execution happens back to the replica to be applied to
the canister's state.

Currently, the checks are somewhat spread out. The sandbox as well as
various execution algorithms (call_or_task, response, install_code)
checked whether the execution was successful or not to attempt to apply
changes. Canister logs were also handled in a special way being passed
along with other information from the sandbox as they have different
semantics than most other changes (specifically, they should persist
even in case of traps).

The PR attempts to consolidate all logic and decision making in one
single place. The system API becomes essentially that place and the main
function to look for is `take_system_state_modifications`. This function
captures all the business logic of which changes should be returned and
are relevant depending on the execution context as well the execution
result. The sandbox is then converted to simply pass over the changes as
needed back to the replica while the replica is simplified to avoid
checking the execution result and applying changes that it was given (of
course retaining the sanity checks that were happening when applying the
changes in case the sandbox execution is compromised).

Finally, logs are also incorporated in the changes returned back by the
sandbox and this allows us to remove a bunch of places spread across
embedders and execution where we would append the log after the end of
an execution.

The main benefits of this refactoring should be:
1. Easier to inspect and verify what kind of changes need to be
persisted (you only need to look at one place). Previously, it was very
easy to forget a place where you need to perform a check or even think
you're missing a check only to realize it's already happening in another
level.
2. Maintainability in the future should be enhanced: in case we need to
add a new field in the system state, the compiler will guide you to
update the big match statement in `take_system_modifications` forcing
the writer to think what they should do with the new field.