fix(ckbtc): use scope guard to prevent double minting (#3930)
XC-254

The loop over processable_utxos may keep looping and eventually panic
when exceeding the instruction limit if there is a large number of utxos
and all of them have a value less than check_fee. If it does panic, then
any previous state modification will not be persisted which spells
trouble if a previous utxo has already been minted. Although this is a
very unlikely scenario, it does still present a non-zero risk.

The fix is to use a scope guard that will be triggered under this
situation and update the utxo to a special status to prevent it from
being minted again.