chore(IDX): clean up status
This removes (more) unused status env vars, and removes the workaround
used to clean up the volatile status by instead using stable status for
farm metadata.

This means farm tests will be retriggered if the jobs are run under a
new CI job name, which is actually already the case since ic-os images
are not cached (we rely on `diff.sh` for skipping tests).