chore(crypto): CRP-2682 Change key derivation used in VetKD to BLS12-381 hash_to_scalar (#3736)
The approach originally used for the vetKD prototype for key derivation
used SHAKE, largely as an artifact that we implemented an IBE demo as
part of the same prototype, and SHAKE happened to be convenient for
implementing IBE. However for production use of vetKD we will switch to
using `hash_to_scalar`, a variant of RFC 9380's `hash_to_field`, since
we already depend on the security of `hash_to_field` for implementing
the hash to curve operation, upon which the security of vetKD depends.
Thus we slightly reduce the number of cryptographic primitives vetKD
depends on for its security.