feat: list topics (#3394)
This PR introduces the concept of topics to the SNS through the
`list_topics` api. The intention is that UIs to the NNS can query
`list_topics` and use the result to render a list of all available
topics (as well as the functions within them).

When registering a GenericNervousSystemFunction, a topic can also be
specified, and it will appear under that topic under the output of
`list_topics`.

We want to implement following on topics at some point in the future,
but that will be done in a separate PR.