fix: Use correct registry version to validate dealings inside of block payloads (#3944)
We should use the same registry version when validating NiDKG dealings,
regardless of whether they come from a consensus block payload or the
DKG pool.

This PR changes the validation of dealings that are part of a block
payload to use the registry version of the dealing's config. This is
consistent with the registry version that is used to
[create](https://sourcegraph.com/github.com/dfinity/ic/-/blob/rs/consensus/dkg/src/lib.rs?L157)
and
[validate](https://sourcegraph.com/github.com/dfinity/ic/-/blob/rs/consensus/dkg/src/lib.rs?L253)
the dealing as part of the DKG client.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>