feat(IC-1579): allow disabling the TLA instrumentation even with test feature (#3958)
When the test feature was on, we have so far always recorded TLA traces
in governance. But this can be problematic when a canister is built with
this feature turned on, because the traces would be collected in the
background. This is a problem for canisters with large states (since the
traces keep multiple copies of the state in the Wasm heap), but even
with moderate state sizes it would also effectively leak the canister
memory over time, since the traces were never emptied.

Change the instrumentation infrastructure to not record any traces when
both:
1. a `TLA_TRACES_LKEY` task-local (`LocalKey`) isn't set (this is meant
to be used from pure Rust tests and is set by the `with_tla_trace_check`
attribute macro)
2. `TLA_TRACES_MUTEX` isn't set (this is currently just disabled in
governance, but could be used to collect traces from a canister, but the
infrastructure is lacking)