feat: CON-1346 CON-1407 Don't use `rsync` to copy the state during recoveries (#3853)
https://github.com/dfinity/ic/pull/2948 introduced a change which avoids
using `rsync` during the "upload state" step of recoveries performed on
the replica node directly. This PR does the same for the "download
state" step by using `cp` instead.

Additionally, we add a parameter `--wait-for-cup-node` which is used to
specify the IP of the node to be used for polling the current recovery
CUP. This is necessary if no IP was specified in either the "download"
or the "upload" step.

Futhermore, we introduce a new function `confirm_exec_cmd`, which will
only execute the given command after explicit user confirmation (similar
to how it is already done for `rsync` and `ssh` commands). This allows
for easier sanity checks to improve the confidence in the new process.

Finally, we add a system test for local node recoveries, which performs
a recovery by sending a single `ic-recovery` to a subnet node via `ssh`.
Node that, for now, we need to skip the download & merge certification
steps, as these require ssh agent forwarding, which seems difficult to
do in the test environment.

The new system test exercises the recovery variant without version
upgrade, failover nodes, or chain key resharing.