test(sns): Porting sns-testing to the ICP mono repo (#3979)
This PR implements the next-gen
[sns-testing](https://github.com/dfinity/sns-testing) tool. It can now
be written in Rust and continuously tested against the latest SNS
features.

At the core of this tool lies the [PocketIC
framework](https://crates.io/crates/pocket-ic/6.0.0), the testing
framework capable of creating a fully deterministic environment for
testing complex ICP dapp scenarios.

This PR is authored by Serokell and reviewed by the DFINITY Foundation.
Please refer to the reviews here:
https://github.com/dfinity/ic/pull/3654

---------

Co-authored-by: Roman Melnikov <roman.melnikov@serokell.io>
Co-authored-by: Roman Melnikov <rvembox@gmail.com>