chore(IDX): Simplify versioning (#3984)
This simplifies the mechanism used to stamp the released artifacts with
a version.

In the non-release mode, a constant version is used. In release mode
(with `--config=stamp`) the HEAD commit id and timestamp are used.

This also removes some workarounds that were necessary to re-trigger
bazel evaluation (timestamp file, etc) as well as the distinction
between "IC version" and "IC RC-only version".

Moreover the upload check (to ensure we don't upload non-release
artifacts) is removed because (1) uploads are (currently) always
necessary (for SHA256SUMS files) and (2) the check was using an
incorrect string in the comparison (couple extra zeros) which meant it
never triggers.