test: CON-1422 vetKD payload builder/verifier unit tests (#3886)
This PR adds unit tests for the new vetKD payload builder. Some unit
tests are still annotated with `#[should_panic(expected = "not yet
implemented")]`, as the current implementation doesn't call the correct
crypto endpoints to create and validate combined vetKD shares yet. This
will be added in a subsequent PR, at which point the annotation will be
removed.