refactor(EXC-1905): Take cycles from call context when a refund is to be returned (#3985)
This change makes it more clear that whenever a refund needs to be
computed when processing a result of a canister execution, the available
cycles in the call context would need to be emptied. This is important
to maintain spec compliance and disallow accepting cycles after the call
has been replied to.

In the current version of the code, this was not happening in the call
context manager and instead we kinda relied on some logic in the system
API that would return 0 cycles can be accepted when the call was already
responded (without looking anything up in the call context). This is a
bit error prone as two different components are attempting to uphold a
property. In fact, we had filed a bug ticket thinking that we are not
properly maintaining spec compliance in this case. This was incorrect
but it shows that the current version of the code is hard to verify
(especially if one only looks at one place, namely the call context
manager).

With the changes here the logic of clearing the available cycles remains
encapsulated in the call context manager module and the system api will
always look up the cycles in the call context without any special cases.