refactor: Use two instances of the query service for queries and https outcall transform functions (#3992)
In a previous
[change](https://github.com/dfinity/ic/commit/3760f0a2898a1ff66a22b14a626666ce481be9f8),
the older `AnonymousQueryService` was dropped in order to re-use the
same `QueryService` for handling both regular queries and transform
functions from https outcalls.

While this change is a step in a good direction as we don't have mostly
duplicate code in `AnonymousQueryService`, it couples regular queries
with transform function executions. To provide better quality of
service, go back to using two instances of the same service. This way we
can get the best of both worlds: code reusability as well as isolation
between the two types of queries.