chore(EXC-1841): Remove old pmap canister (#4002)
This is a very old canister using quite unsafe operations and it was
only added as part of the old COW implementation of Orthogonal
Persistence that we have long abandoned (and was never shipped to
production). Instead of fixing the TODO (that would basically require
switching the canister to use a saner approach wrt handling its state),
remove the canister as it doesn't seem to be used anywhere else.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>