test: Make tests not squelch color when being run by bazel. (#4006)
What this literally does is sets the `CLICOLOR_FORCE` environment
variable to `true`. This causes terminal output to be colored, if the
test follows [that convention]. In particular, the Rust `colored`
library uses that convention. In particular, canbench uses `colored`.
This makes the output of canbench easier for humans to consume (and we
should really care about humans, such as ourselves).

[that convention]: http://bixense.com/clicolors/