refactor: Remove install_code from production code and move it to the tests that use it (#4015)
`CanisterManager::install_code` is the legacy function that was handling
`install_code` requests as mentioned in the doc comment for it. It's
only now used in `canister_manager` tests. (Most of) These tests use
their own setup function and not the expected `ExecutionTest` framework.
Moving the tests to use it is a big effort but it would unify things as
these tests would naturally go through the production code paths
(instead of maintaining a hidden method that delegates to
`install_code_dts`).

As a first step towards cleaning this up, moving the logic to the tests
module that actually uses it still. One test is also migrated as it
wouldn't be handled properly with the changes (the checks on canister's
existence happen one step earlier in the production code that uses
`install_code_dts`, one more reason to unify how these tests work).