feat: CON-1422 Call crypto in vetKD payload builder (#3905)
This PR adds the missing crypto calls to create and validate combined
vetKD key shares.

When combining shares, we ignore `UnsatisfiedReconstructionThreshold`
errors. Other errors trigger a warning and metric increase.

When verifying combined keys as part of a block, we invalidate the block
on "reproducible" errors. Non-reproducible errors are retried at a later
point.