chore: Allow any order of change actions in `purging_non_finalized_blocks_test` (#4033)
This test's expectation has to be adapted whenever the order of the
block proposals returned by the purger changes. This usually happens
when the block payload struct is updated in a way that changes the
lexicographic order of the block hashes. With this PR we allow any order
of the returned change actions.