feat(CON-1419): Deliver VetKeys to execution (#3860)
This PR adds a function, which returns the `MasterPublicKeys` and
`NiDkgId` corresponding to the `MasterPublicKeyId`s that are active
VetKD keys in the subnet.

The `MasterPublicKey`s are then delivered to execution. This enabled the
`vetkd_public_key` function to work. The corresponding system test is
adapted to demonstrate that.

The `NiDkgId`s are not delivered to execution yet (this is captured in
CON-1420).
Also, currently there is no unit test to cover the functionality.

---------

Co-authored-by: Leo Eichhorn <99166915+eichhorl@users.noreply.github.com>