feat(node): Move hash resolution into bazel (#3427)
NODE-1431
System tests that depend on IC OS images will now receive a URL and hash
via environment variables. Previously, only a URL was supplied, and URL
to hash resolution was handled in Rust code. This moves the same
resolution into bazel, which allows us to supply arbitrary URLs and
hashes to the test binary.

---------

Co-authored-by: Andrew Battat <andrew.battat@dfinity.org>