fix(icp-rosetta): [FI-1649] support refresh voting power with hot key (#4050)
Callers should be able to refresh voting power in Rosetta ICP while
signing with hotkeys, but currently Rosetta derives the neuron
subaccount from the signer key, so if the signature is from the hotkey
rather than the neuron owner, Rosetta is not able to derive the real
neuron subaccount.

This PR adds the `controller` field to the refresh voting power
operation metadata, with which a hotkey signer can specify the real
controller of a neuron so that Rosetta can correctly derive its
subaccount.

This was successfully tested using real neurons in mainnet.