feat(ICRC_Ledger): FI-1664: Forbid setting interpreted ICRC ledger metadata (#3767)
Disallow setting the following metadata in the initialization and
upgrade arguments of the ICRC ledger:

- `icrc1:symbol`
- `icrc1:name`
- `icrc1:decimals`
- `icrc1:fee`
- `icrc1:max_memo_length`

These fields are already set explicitly in the arguments, from there the
ledger reads them and interprets them. Setting them in the vector of
metadata is confusing, since they are then returned in the response to
`icrc1_metadata` queries, but they have no effect on the operation of
the ledger.