chore(consensus): change the order of random beacon (share) validations (#1918)
There are a bunch of `Invalid consensus artifact (The parent hash of the
beacon was not correct)` warnings in logs. I took a closer look and in
all the cases the running replica version was different than the replica
version under which these artifacts were produced. If I understand
correctly this happens when some replicas upgrade faster than the rest
and receive (discarded) random beacons from slow peers still running the
old replica version.

I think the log message is a bit misleading, so I changed the order of
validations of random beacons and their shares - we first check the
signatures and replica versions and only then we check that the parent
hash is valid. If there is a replica version mismatch, we won't see any
message in our logs