chore(IDX): standardize token usage (#4074)
This PR makes two changes:
- remove workflow permissions where a GitHub token is used (the workflow
permissions are not required and only cause confusion)
- standardize how we reference the github token -> we specify it during
the checkout step so it is automatically used for the rest of the
workflow. It is not required to re-specify the token using the
`GH_TOKEN` env var in later steps

Tested all the workflows and they work, including the container image
update workflow.

---------

Co-authored-by: IDX GitHub Automation <>