chore(IDX): Don't upload artifacts on non-release
This changes the build determinism check to not rely on the shasums to
be uploaded to S3.

Instead, all SHA256SUMS generated during the build are aggregated and
uploaded as a GHA artifact. A new config is introduced -- `release --
which turns on stamping (`--config=stamped`) and enables upload to S3.

Note: the implementation currently just checks all built files and grabs
the ones called `SHA256SUMS`, which might break in the future if the
build changes; but has the benefit of being very simple.