refactor(nns): Move seeding to timer task (#4123)
The `timer_task` library defines some common patterns to schedule
timers, and metrics collection will be added to the library. Gradually,
the existing tasks in the NNS Governance should be changed to scheduling
with `timer_tasks`. This PR should be a refactoring, since the
`timer_task` library is modeled after the existing way the NNS
Governance schedule timers.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>