feat(crypto): CRP-2699 adapt vetKD API to use derivation domain blob instead of path (#4049)
Changes the hierarchical vetKD `derivation_path: vec blob` to a
non-hierarchical `derivation_domain: blob` because the vetKD public key
derivation algorithm is actually non-hierarchical. This simplifies the
API. Callers can still embed hierarchical structure into the derivation
domain blob if they wish.

The respective change in the IC spec PR is
https://github.com/dfinity/portal/pull/3763/commits/3bbb8f4f5de2bd579c81f85ed2dbff67322164c9.

The name _derivation domain_ is not set in stone and is up for
discussion. One possible alternative is _derivation context_, but there
may be others.

Note that the `SignWithThresholdContext` currently contains a field
`SignWithThresholdContext::derivation_path: Vec<Vec<u8>>` that currently
applies no matter if the context relates to ECDSA, Schnorr, or VetKD
(based on the `SignWithThresholdContext:args: ThresholdArguments`). Now
that the VetKD variant uses just a blob instead of a vec of blobs, the
derivation path is flattened to form the derivation domain. In a later
step, we may consider moving `SignWithThresholdContext::derivation_path`
into `{Ecdsa|Schnorr}Arguments` and add
`VetKdArguments::derivation_domain`, given that not all three arguments
don't use the hierarchical derivation path any more. This, however, is
not strictly necessary and rather a cleanup action, which would require
a multi-step migration because the `SignWithThresholdContext` is
persisted in the state.