fix: Compatibility xnet test (#4122)
This
[commit](https://github.com/dfinity/ic/commit/a944a8b331f1562508a5dde68f6eafcfbb54fe44)
broke the xnet compatibility test by adding a new field to
`CanisterStatusResultV2`. This is an artifact of depending on the
private management types instead of using the public version of the
types as client code is supposed to do. Replicas never attempt to decode
this type, they only produce it so there's no real issue with the IC's
compatibility.

The quick fix here is to make a temporary copy of the type without the
new field until the change is rolled out and then drop the copy in
favour of using the public types. Additionally, the
[check](https://sourcegraph.com/github.com/dfinity/ic@6c12598a8e5ab3a1a241af5b51d5a27990b8177e/-/blob/rs/rust_canisters/canister_test/src/canister.rs?L810-L820)
of `canister_status` can be removed altogether when stopping a canister
as it's redundant -- if the `canister_stop` request receives a reply,
then the canister is stopped (otherwise a timeout error would be
received).

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>