fix: EXC-1946: Fix `InstanceStats` metrics (#4062)
Currently, metrics are not reported for DTS executions, also when heap
dirty pages exceed 1 GiB.

This PR moves `InstanceStats` metrics from the hypervisor down to the
sandbox execution level. That's why all the prefixes changed from
`hypervisor_` to `sandboxed_execution_`. The descriptions and properties
are otherwise the same.

The Rust formatter was having trouble with long lines in the metrics, so
those are also fixed and formatted now.

This fixes RUN-982