refactor: Convert parts of the canister_test framework to use public management canister types (#4135)
This PR converts parts of the canister_test framework to depend on the
public management types instead of the private ones used by the replica.
This allows for better isolation between making changes to the replica
and having the client evolve at its own pace when it's ready to pick up
the changes.

Some types are still imported from the private crate as it's a much
bigger effort to move that has inter-dependencies (e.g. the state
machine framework needs to be adapted or pocket ic should be fully used
for canisters tests as another way to break the dependency). This part
is considered a future improvement and thus is not included here.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>