fix(icp-rosetta): remove deadlock in rosetta blocks (#4147)
`self.ledger.read_blocks()` returns a `Box<RwLock>` that has a RW mutex
to the blocks database connection.

In the current implementation, there are at least a couple of situations
in which we get a read lock and then call another function that also
gets a read-lock for themselves before the caller released theirs. Most
of the time this is ok (since it's a RwLock) but in some rare scenarios
a different thread might try to get a write lock in between the two read
locks from the first thread, leading to a deadlock (the first read lock
won't be released until the second read lock is released, but the second
read lock is waiting on the write lock to be released, which is itself
waiting for the first read lock to be released).

As far as we know this is only affecting the rosetta blocks
implementation, which is not in prod, but there's nothing preventing the
same conditions from happening elsewhere and just not being caught by
tests at the moment.

This PR simply remediates the rosetta block tests flakiness by releasing
the top read locks (by making the returned box go out of scope) when
they're not needed, but a more thorough work is being done on fixing the
confusing lock handling we have now.

I've run the flaky tests 300x now and didn't get any failure, as opposed
to before when ~80 runs were enough to reproduce the issue.