feat: [EXC-1940] Publish ic-error-types (#4120)
This PR moves `/rs/types/error_types` to `/packages/ic-error-types`. 

The dependents now point to the new source by path. This will change
once the new crate is published to crates.io, from which point on they
have a versioned dependency and can choose to upgrade independently of
the replica's implementation.

Not all code can just be moved to packages/, because code in /packages
cannot not depend on code in /rs. Therefore,
- two traits have been copied (`include!()` did not work with bazel,
only with cargo)
- some `From`/`Into` impl blocks have been moved to the `protobuf`
crate, because Rust's orphan rule demands they are implemented in the
same crate as either the trait (which we have no control over) or the
type. The public ic-error-types crate has no access to the protobuf type
counterparts, so it must be the protobuf types that are aware of the
public types instead.