chore: Let PocketIC library depend on public ic-management-canister-types (#3986)
This is an imperfect step in the right direction. 

Follow-ups and explanations: 
- ~The [public types have a
typo](https://github.com/dfinity/cdk-rs/issues/558) in a field name
which will lead to breakage when it is fixed.~ EDIT: Fixed with version
0.0.2
- Some of the names of mgmt types that the PocketIC library defined in
its `management_canister` module are different from the ones in the
public crate. For now, they are imported `as` their old name, in order
to reduce changes in this PR. They should probably be consistently
renamed. Depending on whether these types are exposed to PocketIC users,
this might be a breaking change, hence this PR does not perform this
renaming step.
- The module `management_canister` still contains a few types, because
they are not defined YET in the current version of the public crate, ~or
they are more recent versions of those types (e.g., with an additional
field)~. As the public crate is upgraded and contains all necessary
types, this module will hopefully disappear, and all necessary
management canister types can be imported via the public crate. For
experimental pre-releases, the `ic-management-canister-types` crate and
the `pocket-ic` crate have to both publish beta-releases on crates.io.

This PR changes more crates in `ic/rs/...` than it should, because 
- These crates depended on pocket-ic (more or less directly) for certain
management canister types. Now that PocketIC does not define these
itself anymore, all these consumers need to depend on the real
dependency, some with renaming `as` the old name.

### To the codeowners of these crates: 
If you have only imported PocketIC for the management canister types,
please remove the dependency and depend on the public crate. If you find
yourself depending on ic-management-canister-types-**private** or
`"//rs/types/management_canister_types"`, please also move to [the
public crate](https://crates.io/crates/ic-management-canister-types).

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>
Co-authored-by: IDX GitLab Automation <idx@dfinity.org>